Added Encryption
The app now have the ability to work in both encrypted mode and in encrypted mode. The only difference is, that in the encrypted mode, the questions and beeps (answers to the questions) are encrypted. 

The password for encryption used is not the same as for encrypting files. It uses the password, that the user uses to login. The password is not saved, what is saved in the SHA256 hash of the password. When login, the saved hash is compared to the hash of the password inputted. 

In order to make the app work, I also needed to do the following things:
* I needed to make sure that the exceptions for when the data does not exist are handled alright
* I changed the way the answers for one single beep are collected together
* I changed all componentWillMount to componentDidMount because of this: https://reactjs.org/docs/react-component.html#unsafe_componentwillmount

+++ src/DeleteFile.js
@@ -10,11 +10,28 @@ class ReadFromFile extends Component {
         title: "Delete File",
     };
 
-    state = {};
+    state = {
+        currentStatus: "",
+    };
 
     deleteFile = () => {
         const path = RNFS.DocumentDirectoryPath + "/test.txt";
-        RNFS.unlink(path);
+
+        RNFS.exists(path).then(result => {
+            if (result) {
+                RNFS.unlink(path)
+                    .then(value =>
+                        this.setState({
+                            currentStatus: "Datoteka je bila izbrisana",
+                        }),
+                    )
+                    .catch(err =>
+                        this.setState({
+                            currentStatus: "Ta datoteka ne obstaja",
+                        }),
+                    );
+            }
+        });
     };
 
     render() {
@@ -25,6 +42,7 @@ class ReadFromFile extends Component {
                     onPress={() => this.deleteFile()}>
                     <Text>Delete File</Text>
                 </TouchableHighlight>
+                <Text>{JSON.stringify(this.state.currentStatus)}</Text>
             </View>
         );
     }
